+++ inst/plotlyjs.R
@@ -26,6 +26,12 @@ file.copy(
   file.path("inst/htmlwidgets/lib/plotlyjs/locales", sub("^plotly-locale-", "", basename(locales))),
   overwrite = TRUE
 )
+# update typed array polyfill
+download.file(
+  "https://raw.githubusercontent.com/plotly/plotly.js/master/dist/extras/typedarray.min.js",
+  "inst/htmlwidgets/lib/typedarray/typedarray.min.js"
+)
+
 # update the plot schema
 Schema <- jsonlite::fromJSON(Sys.glob("*plotly.js*/dist/plot-schema.json"))
 usethis::use_data(Schema, overwrite = T, internal = T)