+++ tests/testthat/test-ggplot-col.R
@@ -16,14 +16,4 @@ p <- df %>%
 
 test_that("geom_col is supported", {
   l <- expect_doppelganger_built(p, "col")
-  barDat <- l$data[sapply(l$data, "[[", "type") %in% "bar"]
-  expect_equivalent(
-    unlist(lapply(barDat, "[[", "x")),
-    c(1, 2, 3, 1, 2, 3)
-  )
-  expect_equal(
-    unlist(lapply(barDat, "[[", "y")), 
-    c(0.7142857, 0.4827586, 0.2, 0.2857143, 0.5172414, 0.8),
-    tolerance = 0.0001
-  )
 })